requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
Django==1.11.20
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-filter==2.0.0
djangorestframework==3.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
lxml==4.3.2
Markdown==3.0.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-docx==0.8.10
pytz==2018.9
PyYAML==5.1
requests==2.21.0
uritemplate==3.0.0
urllib3==1.24.1
